src/image_sorter/requirements.txt
./libs/filemanager
./libs/lifecyclemanager
./libs/messagebroker
nats-py
python-dotenv
numpy
pillow